Fix numba compilation setup?

+++ asv_bench/benchmarks/reduce.py
@@ -20,14 +20,14 @@ class ChunkReduce:
         if "numba" in engines:
             for func in funcs:
                 npg.aggregate_numba.aggregate(
-                    np.ones((100,), dtype=int), np.ones((100,), dtype=int), func=func
+                    np.ones((100,), dtype=int), np.ones((100,)), func=func
                 )
         if "numbagg" in engines:
             for func in funcs:
                 try:
                     method = getattr(flox.aggregate_numbagg, func)
-                    method(np.ones((10,), dtype=int), np.ones((10,), dtype=int))
-                    method(np.ones((10,), dtype=int), np.ones((10, 10), dtype=int))
+                    method(np.ones((10,), dtype=int), np.ones((10,)))
+                    method(np.ones((10,), dtype=int), np.ones((10, 10)))
                 except AttributeError:
                     pass
